Bugfix/raspbian compilation (#180)
* Replace CLI library with custom argparser for micro-ros-agent Foxy

* Fix compilation issues

* Refs #9119. Fix agent compilation for raspbian

* CLI profile must remain ON for develop branch
