+++ include/uxr/agent/utils/ArgumentParser.hpp
@@ -798,71 +798,6 @@ private:
     std::unique_ptr<AgentType> agent_server_;
 };
 
-namespace utils
-{
-
-bool usage(
-        bool no_help)
-{
-    std::stringstream ss;
-    ss << "Usage: 'MicroXRCEAgent <udp4|udp6|tcp4|tpc6";
-#ifndef _WIN32
-    ss << "|serial|pseudoterminal";
-#endif // _WIN32
-    ss << "> <<args>>'" << std::endl;
-    if (no_help)
-    {
-        ss << "For a more detailed description about all the available arguments, ";
-        ss << "please execute the agent with '-h/--help' option." << std::endl;
-    }
-    std::cout << ss.str();
-    return false;
-}
-
-TransportKind check_transport(
-        const std::string& transport)
-{
-    const std::unordered_map<std::string, TransportKind> valid_transports = {
-    {"udp4", TransportKind::UDP4},
-    {"udp6", TransportKind::UDP6},
-    {"tcp4", TransportKind::TCP4},
-    {"tcp6", TransportKind::TCP6},
-#ifndef _WIN32
-    {"serial", TransportKind::SERIAL},
-    {"pseudoterminal", TransportKind::PSEUDOTERMINAL},
-#endif // _WIN32
-    };
-
-    if (valid_transports.find(transport) != valid_transports.end())
-    {
-        return valid_transports.at(transport);
-    }
-    return TransportKind::INVALID;
-}
-
-Middleware::Kind get_mw_kind(
-        const std::string& kind)
-{
-#ifdef UAGENT_FAST_PROFILE
-    if ("dds" == kind)
-    {
-        return eprosima::uxr::Middleware::Kind::FASTDDS;
-    }
-    if ("rtps" == kind)
-    {
-        return eprosima::uxr::Middleware::Kind::FASTRTPS;
-    }
-#endif
-#ifdef UAGENT_CED_PROFILE
-    if ("ced" == kind)
-    {
-        return eprosima::uxr::Middleware::Kind::CED;
-    }
-#endif
-    return eprosima::uxr::Middleware::Kind::NONE;
-}
-
-} // namespace utils
 } // namespace parser
 
 /*************************************************************************************************
